bug fixing

+++ src/main/java/org/webcastellum/ResponseFilterWriter.java
@@ -174,7 +174,7 @@ public final class ResponseFilterWriter extends AbstractRelaxingHtmlParserWriter
         this.honeylinkSuffix = honeylinkSuffix;
         this.honeylinkMaxPerPage = honeylinkMaxPerPage;
         if (honeylinkMaxPerPage > 0) {
-            this.honeylinkRandom = randomizeHoneylinksOnEveryRequest ? null : new Random( this.servletPath.hashCode()+WebCastellumFilter.customerIdentifier );
+            this.honeylinkRandom = randomizeHoneylinksOnEveryRequest ? null : new Random( ((long) this.servletPath.hashCode())+WebCastellumFilter.customerIdentifier );
             this.tagPartCounterTarget = HoneylinkUtils.nextTagPartCounterTarget(this.honeylinkRandom);
         } else this.honeylinkRandom = null;
     }
